refactor: update detail page with new api
