--- src/components/pages/characterPage/characterPage.css
@@ -1,9 +1,0 @@
-.character-page {
-  min-width: 350px;
-  background-color: rgba(0, 0, 0, 0.219);
-}
-
-.character-page__close {
-  margin: 2rem;
-  background-color: red;
-}
