+++ src/components/pages/productPage/productPage.css
@@ -1,0 +1,14 @@
+.product-page {
+  width: 35rem;
+  padding: 1rem;
+  background-color: rgba(0, 0, 0, 0.219);
+}
+
+.product-page__close {
+  margin: 2rem;
+  background-color: red;
+}
+
+.product-page__img {
+  max-width: 100%;
+}
